Mentorxhub/requirements.txt
asgiref==3.11.0
Django==6.0
pillow==12.0.0
sqlparse==0.5.3
django-allauth==65.13.1
requests==2.32.5
PyJWT==2.10.1
cryptography==46.0.3
django-crispy-forms
crispy-bootstrap5